Add serial support to kernel
The kernel previously wasn't ready to accept incoming serial connection
through e.g. the hypervisor and instead relied solely on the VGA output.
This is a bit stupid - as the VGA out also just displays the console.

Luckily linux has support for console/serial out and that can be
employed with libvirt through various ways. So enabling the relevant
kernel flags, and this works.

Signed-off-by: Sophie 'Tyalie' Friedrich <dev@flowerpot.me>