Fix: non existing gpg tmp dir during ssh login
= Issue
During initial login over SSH, the RemoteForward of the gnupg agent
socket did not work properly or rather errored out, because it could not
create the socket. The underlying cause of the problem was that the
directory `/run/user/$uid/gnupg` would not exist and SSH using
RemoteForward is unwilling to create directories while trying to forward
unix sockets.

Normally the directory would be created by `gpgconf --create-socketdir`
during start of the gpg-agent, which does not happen in our case.

= Solution
Using the solution provided here solved the issue.
https://dev.gnupg.org/D435

As the directory /run/user/$uid only exists due to systemd, it is almost
always applicable. The solution in the gnu docs that suggests adding a
line to the .bashrc is far to late in the login process.

Signed-off-by: Sophie 'Tyalie' Friedrich <dev@flowerpot.me>